split CategoricalDisplay to many subcomponents
