+++ src/components/categoricaldisplay/index.js
@@ -1,0 +1,13 @@
+import CategoricalDisplay from './CategoricalDisplay.jsx';
+import { FilterPanel } from './FilterPanel.jsx';
+import { NormalItem, ActiveItem, BlankItem } from './Item.jsx';
+import { ResultsPanel } from './ResultsPanel.jsx';
+
+export {
+    CategoricalDisplay,
+    FilterPanel,
+    NormalItem,
+    ActiveItem,
+    ResultsPanel,
+    BlankItem,
+};
